# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/item/createItem": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "itemName": {
                  "example": "any"
                },
                "itemCode": {
                  "example": "any"
                },
                "itemType": {
                  "example": "any"
                },
                "sku": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/item/getItems": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/item/deleteItem/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/item/updateItem/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "itemName": {
                  "example": "any"
                },
                "itemCode": {
                  "example": "any"
                },
                "itemType": {
                  "example": "any"
                },
                "sku": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  }
}